ユーザー側と管理者側の分離
garageフォルダを作成し、ユーザー側のファイル群を格納
