--- css/style.css
@@ -1,4 +1,0 @@
-.basic, .simple{
-    border: solid 1px black;
-    cursor: pointer;
-}
