+++ garage/css/style.css
@@ -1,0 +1,4 @@
+.basic, .simple{
+    border: solid 1px black;
+    cursor: pointer;
+}
