+++ garage/tpl/header/header.php
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html lang="ja">
+<head>
+    <meta charset="UTF-8">
+    <title>GARAGE（仮）</title>
+    <meta name="viewport" content="viewport-fit=cover,user-scalable=no">
+    <meta name="apple-mobile-web-app-capable" content="yes">
+    <meta name="apple-mobile-web-app-status-bar-style" content="black-translucent">
+    <meta name="apple-mobile-web-app-title" content="テストサイト">
+    <link rel="apple-touch-icon" sizes="180x180" href="./images/howl.jpg">
+
+    <link rel="stylesheet" href="./css/reset.css">
+    <link rel="stylesheet" href="./css/header.css">
+    <link rel="stylesheet" href="./css/footer.css">
+    <link rel="stylesheet" href="./css/style.css">
+
+    <link href="https://fonts.googleapis.com/css?family=Nunito&display=swap" rel="stylesheet">
+</head>
+<body>
+    <header class="basicOnly">
+        <h1><img src="./images/logo.jpg" alt="GARAGE（仮）" class="basic toToppage"></h1>
+        <div id="toNavi"><img src="./images/" alt="ナビ表示用"></div>
+    </header>
