+++ garage/tpl/main/policy.php
