+++ garage/tpl/navi/navi.php
@@ -1,0 +1,21 @@
+    <nav class="basicOnly">
+        <dl>
+            <dt><img src="images/users/00001/account.jpg" alt=""></dt>
+            <dd>＄ユーザーネーム</dd>
+            <dd>＄ユーザーのメアド</dd>
+        </dl>
+        <hr>
+        <ul>
+            <li class="simple toRegist_1">新規登録</a></li>
+            <li class="simple toLogin">ログイン</a></li>
+            <li class="basic toChange">ユーザー情報確認/変更</li>
+            <li class="basic toReserv_list">予約一覧</li>
+            <li class="basic toNotice">お知らせ</li>
+            <li class="basic toHelp">ヘルプ</li>
+            <li class="basic toTerm">利用規約</li>
+            <li class="basic toPolicy">プライバシーポリシー</li>
+            <li class="simple toLogout">ログアウト</a></li>
+            <li class="simple toWithdraw">退会</a></li>
+        </ul>
+    </nav>
+    <section id="ajaxArea">
