--- tpl/footer/footer.php
@@ -1,15 +1,0 @@
-        </section><!-- #innerAjaxArea -->
-        <footer class="basicOnly">
-            <ul>
-                <li><img src="./images/garage/payment.jpg" alt="予約一覧画面アイコン" class="basic toReserv_list"></li>
-                <li><img src="./images/garage/schedule.jpg" alt="トップページアイコン" class="basic toToppage"></li>
-                <li><img src="./images/garage/reserv.jpg" alt="予約画面アイコン" class="basic toReserv"></li>
-            </ul>
-        </footer>
-    </div><!-- #outerAjaxArea -->
-
-    <script src="./js/jquery-3.3.1.min.js"></script>
-    <script src="./js/function.js" type="text/javascript"></script>
-    <script src="./js/public.js" type="text/javascript"></script>
-</body>
-</html>
