--- tpl/main/regist_complete.php
@@ -1,10 +1,0 @@
-<h1><img src="./images/garage/simple_logo.jpg" alt="ロゴマーク"></h1>
-<h2>登録完了</h2>
-<div>
-    <p>登録完了メールを<br>
-    お客様のメールアドレス宛に<br>
-    送信しました。</p>
-    <p>ご確認ください</p>
-</div>
-<button type="button" class="basic toToppage">TOPページへ</button>
-<button type="button" class="back">引き続きお手続きをする</button>
